Fix: defunct processes
When forking processes, we need to make sure that some cleanup is done
for when they complete.  Otherwise, they will become defunct (zombies).

One approach is to wait for them to finish, but we are trying to have
completely async processing, so the alternative is to register a signal
handler for child processes, which is what this commit does.